Move try-related tests to snapstate_try_test.go.
